feat : created BackgroundScheme
