+++ highlight/src/main/java/com/neo/highlight/util/scheme/ColorScheme.java
@@ -7,7 +7,6 @@ import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 
 import com.neo.highlight.core.Scheme;
-import com.neo.highlight.core.ScopeSchemeContract;
 
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -15,35 +14,17 @@ import java.util.List;
 import java.util.regex.Pattern;
 
 /**
+ * Paint the text
  * @author Irineu A. Silva
  */
-public class ColorScheme implements Scheme, ScopeSchemeContract {
-
-    @Nullable
-    private List<Scheme> scopeSchemes;
-
-    @Nullable
-    private final Pattern pattern;
-
-    @ColorInt
-    private final int color;
-
-    private boolean clearOldSpan = false;
+final public class ColorScheme extends BaseColorScheme {
 
     public ColorScheme(@NonNull Pattern pattern, @ColorInt int color) {
-        this.pattern = pattern;
-        this.color = color;
+        super(pattern, color);
     }
 
     public ColorScheme(@ColorInt int color) {
-        this.pattern = null;
-        this.color = color;
-    }
-
-    @Override
-    @Nullable
-    public Pattern getRegex() {
-        return pattern;
+        super(color);
     }
 
     @Override
@@ -51,52 +32,4 @@ public class ColorScheme implements Scheme, ScopeSchemeContract {
     public Object getSpan(@NonNull CharSequence text, int start, int end) {
         return new ForegroundColorSpan(color);
     }
-
-    @Override
-    public boolean getClearOldSpan() {
-        return clearOldSpan;
-    }
-
-    @Override
-    @NonNull
-    public ColorScheme setClearOldSpan(boolean clearOldSpan) {
-        this.clearOldSpan = clearOldSpan;
-        return this;
-    }
-
-    //override ScopeScheme
-
-    @Nullable
-    @Override
-    public List<Scheme> getScopeSchemes() {
-        return scopeSchemes;
-    }
-
-    @Override
-    @NonNull
-    public ColorScheme setScopeSchemes(@Nullable List<Scheme> schemes) {
-        scopeSchemes = schemes;
-        return this;
-    }
-
-    @Override
-    public ColorScheme addScopeScheme(@NonNull Scheme... scheme) {
-
-        if (scopeSchemes == null) {
-            scopeSchemes = new ArrayList<>();
-        }
-
-        scopeSchemes.addAll(Arrays.asList(scheme));
-
-        return this;
-    }
-
-    @Override
-    public ColorScheme clearScopeSchemes() {
-        if (scopeSchemes != null) {
-            scopeSchemes.clear();
-        }
-        return this;
-    }
-
 }
